OUT OF THE LOOP
I know the presentation is already done, but I implemented both SLL and
SLLI within the processor. The SLL code is same as before, so it isnt
just a placeholder anymore, and SLLI uses the opCode previously used for
si. I have also added these instructions to the report capabilities.
This means we need a waveform test for just the SLL commands,
unfortunately.